pyproject.toml
[tool.poetry]
name = "django_blog_api"
version = "0.1.0"
description = ""
authors = ["EkaterinaMoroz <ekaterina-moroz99@rambler.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
asgiref = "3.5.0"
CacheControl = "0.12.11"
cachy = "0.3.0"
certifi = "2021.10.8"
charset-normalizer = "2.0.12"
cleo = "0.8.1"
clikit = "0.6.2"
coreapi = "2.3.3"
coreschema = "0.0.4"
crashtest = "0.3.1"
distlib = "0.3.4"
Django = "4.0.4"
djangorestframework = "3.13.1"
drf-yasg = "1.20.0"
filelock = "3.6.0"
html5lib = "1.1"
idna = "3.3"
importlib-metadata = "4.11.3"
inflection = "0.5.1"
itypes = "1.2.0"
Jinja2 = "3.1.1"
keyring = "23.5.0"
lockfile = "0.12.2"
MarkupSafe = "2.1.1"
msgpack = "1.0.3"
packaging = "20.9"
pastel = "0.2.1"
pexpect = "4.8.0"
pkginfo = "1.8.2"
platformdirs = "2.5.2"
poetry = "1.1.13"
poetry-core = "1.0.8"
psycopg2 = "2.9.3"
ptyprocess = "0.7.0"
pylev = "1.4.0"
pyparsing = "3.0.8"
pytz = "2022.1"
pywin32-ctypes = "0.2.0"
requests = "2.27.1"
requests-toolbelt = "0.9.1"
"ruamel.yaml" = "0.17.21"
"ruamel.yaml.clib" = "0.2.6"
shellingham = "1.4.0"
six = "1.16.0"
sqlparse = "0.4.2"
tomlkit = "0.10.1"
tzdata = "2022.1"
uritemplate = "4.1.1"
urllib3 = "1.26.9"
virtualenv = "20.14.1"
webencodings = "0.5.1"
zipp = "3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
